@ravisantoshgudimetla 2018-05-18
Thanks @godliness for the PR and for the flame graphs. After I looked at them, I decided to run down some more tests to get more data.

Attaching the call graphs with and without cache changes(from this PR, I will add one before and after this PR too). 
- O.S: fedora 27 
- Cores: 8 core
- Golang: 1.10.2
- Source code pulled from latest master branch
- Scheduled: 10k fake pods on 3k fake nodes.

[cpu.pdf](https://github.com/kubernetes/kubernetes/files/2015391/cpu.pdf)
[cpu_withecache.pdf](https://github.com/kubernetes/kubernetes/files/2015392/cpu_withecache.pdf)

I will look more into graphs tomorrow but following is my preliminary analysis:
- While ecache is enabled we are spending around 50.35s (8.52%)of 273.73s (46.30%) in `lookUpResult` fn.
- We are also spending quite a lot of time in gc.
- In total for locking and unlocking we are spending around 47.96sec(for RLock).

Please feel free to add your observations and analysis too.