@godliness 2018-05-18
@ravisantoshgudimetla  Thanks for the detail test information.

From the "cpu_withecache" result:

- Looks like we have already used Rlock in the lookupResult fn

-  runtime.convT2Estring cost time in lookupResult, this is due to we used glog.v(string format) in this fn. and this lead to a lot of time in gc?   

From now on, i just could share this information.