@resouer 2018-05-18
RLock seems solve small part of the problem but not enough. Now I guess the major impact is caused by the lock to protect ecache from race of `factory.go` (need some tests to prove it). If that's the case, we can probably use other approach to replace that lock.