@resouer 2018-05-19
According to my local test, this patch improved operation time a bit:

```
When ecache is enabled:

Original:

BenchmarkScheduling/1000Nodes/1000Pods-32        	     100	  54653471 ns/op	 3915357 B/op	   85775 allocs/op
BenchmarkSchedulingAntiAffinity/500Nodes/5000Pods-32        	     250	  20053089 ns/op	 2846708 B/op	   56706 allocs/op
BenchmarkScheduling/1000Nodes/0Pods-32           	     100	  55889442 ns/op	 3290233 B/op	   73855 allocs/op

With patch:

BenchmarkScheduling/1000Nodes/1000Pods-32        	     100	  50702524 ns/op	 3912885 B/op	   85740 allocs/op
BenchmarkSchedulingAntiAffinity/500Nodes/5000Pods-32        	     250	  18336367 ns/op	 2707880 B/op	   54136 allocs/op
BenchmarkScheduling/1000Nodes/0Pods-32           	     100	  54861638 ns/op	 3224870 B/op	   72515 allocs/op
```