@resouer 2018-05-19
Emm ... op time is increased after patch in your test? 
> Before BenchmarkScheduling/1000Nodes/0Pods-32 1000 5995726 ns/op
After BenchmarkScheduling/1000Nodes/0Pods-32 1000 26993580 ns/op