@godliness 2018-05-19
@resouer cause i added time.StopTimer() in line https://github.com/kubernetes/kubernetes/blob/master/test/integration/scheduler_perf/scheduler_bench_test.go#L160
 I think we don't need calculate the prepare time in the benchmark test, am i right?