@godliness 2018-05-19
@resouer  I tested again in my local test

Because i added b.StopTimer() in line https://github.com/kubernetes/kubernetes/blob/master/test/integration/scheduler_perf/scheduler_bench_test.go#L160
I think we don't need calculate the prepare time in the benchmark test, am i right?

Don't Added b.StopTimer()
```
Before  BenchmarkScheduling/1000Nodes/0Pods-32  1000   29348329 ns/op
After  BenchmarkScheduling/1000Nodes/0Pods-32    1000    26033742 ns/op
```

Added b.StopTimer()
```
Before BenchmarkScheduling/1000Nodes/0Pods-32  1000   11689758 ns/op
After BenchmarkScheduling/1000Nodes/0Pods-32    1000    5951510 ns/op
```

By the way, I need reversed result between "before" and "after", and fill the after result with "Added b.StopTimer()"