@resouer 2018-05-19
Sorry I didn't look it deeply but I'm not quite follow why you need to hack the test right now. While before we can move on, it's important that we should have the same baseline in this discussion ...