@godliness 2018-05-24
@resouer  Please have a look :-)

**Cpus**: 32
**Mem**: 64GB
**Uame -a** :Linux LOCAL 3.10.0-327.28.3.el7.x86_64 #1 SMP Thu Aug 18 19:05:49 UTC 2016 x86_64 x86_64 x86_64 GNU/Linux
**Code Base**: master branch (9c2e51f37c5174d22d0d3819baf5ba4f6a9b2751) with or without this patch
**Golang**: 1.10.2


**Before （Original）**

```
NodeAffinity LabelCount=1
I0524 10:02:25.572008   41645 feature_gate.go:194] feature gates: map[EnableEquivalenceClassCache:true]
I0524 10:02:25.572641   41645 factory.go:1066] Created equivalence class cache
0s    rate: 847    total: 847 (qps frequency: map[847:1])
1s    rate: 599    total: 1446 (qps frequency: map[847:1 599:1])
2s    rate: 524    total: 1970 (qps frequency: map[847:1 599:1 524:1])
3s    rate: 483    total: 2453 (qps frequency: map[847:1 599:1 524:1 483:1])
4s    rate: 476    total: 2929 (qps frequency: map[847:1 599:1 524:1 483:1 476:1])
Scheduled 3000 Pods in 5 seconds (600 per second on average). min QPS was 476
I0524 10:02:33.581535   41645 util.go:83] destroying scheduler
I0524 10:02:33.581607   41645 util.go:87] destroyed scheduler
I0524 10:02:33.581624   41645 util.go:51] destroying API server
I0524 10:02:33.601822   41645 util.go:53] destroyed API server
Minimal observed throughput for 3k pod test: 476
PASS
ok      k8s.io/kubernetes/test/integration/scheduler_perf    11.925s
+++ [0524 10:02:33] ...density tests finished
/export/workspace/src/k8s.io/kubernetes/_output/local/go/src/k8s.io/kubernetes/test/integration/scheduler_perf
+++ [0524 10:02:33] performance test cleanup complete
```

**After (With this patch)**

```
NodeAffinity LabelCount=1
I0524 10:03:44.896844   42385 feature_gate.go:194] feature gates: map[EnableEquivalenceClassCache:true]
I0524 10:03:44.897494   42385 factory.go:1066] Created equivalence class cache
0s    rate: 1416    total: 1416 (qps frequency: map[1416:1])
1s    rate: 1148    total: 2564 (qps frequency: map[1416:1 1148:1])
Scheduled 3000 Pods in 2 seconds (1500 per second on average). min QPS was 1148
I0524 10:03:49.881567   42385 util.go:83] destroying scheduler
I0524 10:03:49.881653   42385 util.go:87] destroyed scheduler
I0524 10:03:49.881678   42385 util.go:51] destroying API server
I0524 10:03:49.883425   42385 util.go:53] destroyed API server
Minimal observed throughput for 3k pod test: 1148
PASS
ok      k8s.io/kubernetes/test/integration/scheduler_perf    8.694s
+++ [0524 10:03:49] ...density tests finished
/export/workspace/src/k8s.io/kubernetes/_output/local/go/src/k8s.io/kubernetes/test/integration/scheduler_perf
+++ [0524 10:03:50] performance test cleanup complete
```

**Without any hack to Benchmark Test**

**Before**

`BenchmarkScheduling/100Nodes/0Pods-32                 1000       5878945 ns/op`

**After**

`BenchmarkScheduling/100Nodes/0Pods-32                 1000       5102698 ns/op`

**With hack to Benchmark Test**

Why i hack the test?

Because we  do not calculate the time of making nodes and destroy them which really cost plenty time on it, and makes the benchmark test not accurate。 If i am wrong, please figure it out. I will really appreciate about that.  

How i hack the test?

Referenced this issue #61582

**Before**

`BenchmarkScheduling/100Nodes/0Pods-32                 1000       2137814 ns/op`

**After** 

`BenchmarkScheduling/100Nodes/0Pods-32                 1000       1219657 ns/op`